Make event handling stateful
This is necessary in various cases. Most immediately, processing the
`Deregister` event requires access to the previous state, because we
are supposed to produce a `FullNodeSetDiff::NodeExit`, which requires
us to know the exit time, which requires us to know the
`exit_escrow_period`, which in turn depends on previous state and
events (since it can be changed causing a contract event to be emitted).

This statefulness may come in handy in other ways, too, for example if
we want to enforce uniqueness of keys in the service rather than rely
on the contract to do it, then event processing will need to know the
full set of keys in the previous state.

This is a pretty big refactor in terms of LOC but it really shouldn't
change much logic, it's just a matter of plumbing.