Fix occasionally hanging test
The issue was that the providers used by the background task use
a caching nonce manager, which doesn't account for failed
transactions. Since the occasional failed transaction is expected
in this test, eventually the local nonce cache diverges from the
actual nonce of the account, and we start building transactions
that never get mined, causing everything to block.

Fixed by using a simple nonce manager that always queries the
nonce from the RPC before building a transaction.